server/schedulers: move evict_leader's tests to the related file
